feat: moved custom golden shadow to tailwind config
In this commit, we have added a custom golden box-shadow to our Tailwind CSS configuration. Also, we updated the hover state styles of a button on the home page. The custom boxShadow has been applied via Tailwind’s utility classes. The previous CSS rules for the boxShadow in `input.css` file have been removed as they are no longer needed.