feat(templates): switch from Django Templating Engine to Jinja2 for improved flexibility and HTMX support
- Switch to Jinja2 from DTE
Jinja2 offers more flexibility and extensibility. Additionally, it allows for a more seamless integration with HTMX, which is critical for our project's front-end.

- Remove django-widget-tweaks
Since django-widget-tweaks is not compatible with Jinja2, it had to be removed to complete the switch. Its functionality has been replaced with custom Jinja2 macros.

- Add jinja2 and django-jinja packages
These packages are essential for integrating Jinja2 with our existing Django project without requiring a complete overhaul.

- Move HTMX-specific templates to a dedicated folder
To make it easier for developers to find and work on HTMX-related templates, we moved them from `<app_name>/partials/` to `<app_name>/htmx_partials/`.

- Create AllAuthExtension for Jinja2
To mimic the django-allauth template tags functionality within Jinja2, we created an AllAuthExtension. This enables us to continue using features like social login buttons with the new templating engine.

- Create render_field Jinja Macro
This macro replicates the functionality of django-widget-tweaks' `render_field` template tag, ensuring we don't lose any existing features with the switch to Jinja2.

- Create render_form_errors Jinja Macro
To make the display of form errors consistent and easier to manage, we encapsulated the logic in a custom Jinja2 macro.

- Add password_reset_done.html Template
To provide better UX, we've added a template that notifies the user when a password reset email has been sent.

BREAKING CHANGE: This commit includes several major changes that affect the project's templating system, as well as directory organization for HTMX templates.